Extracts newly developed tasks into the standard socrata workflow from the dev DAG.
